strict type checks

+++ src/portfolios/domain/positions.service.ts
@@ -38,6 +38,7 @@ export class PositionsService {
     upsertPositionDto: UpsertPositionDto,
   ): Promise<Position> {
     const portfolio = await this.portfoliosRepository.findOne(portfolioUuid);
+    if (!portfolio) throw new NotFoundException('Portfolio not found');
     this.checkOwner(user, portfolio);
     const company = await this.companiesRepository.findBySymbol(
       upsertPositionDto.symbol,
@@ -135,7 +136,9 @@ export class PositionsService {
       positions.map(async (position) => {
         const company = companies.find((c) => c.uuid === position.companyUuid);
         const companyState =
-          await this.companyStatesRepository.getLastByCompanyUuid(company.uuid);
+          await this.companyStatesRepository.getLastByCompanyUuid(
+            company!.uuid,
+          );
         return this.calculatePositionState(position, company, companyState, fx);
       }),
     );
@@ -147,6 +150,7 @@ export class PositionsService {
 
   async deleteByPortfolioUuid(user: User, portfolioUuid: string) {
     const portfolio = await this.portfoliosRepository.findOne(portfolioUuid);
+    if (!portfolio) throw new NotFoundException('Portfolio not found');
     this.checkOwner(user, portfolio);
     const result = await this.repository.deleteByPortfolioUuid(portfolioUuid);
     await this.updatePortfolioState(portfolio);
@@ -160,6 +164,7 @@ export class PositionsService {
   ) {
     const position = await this.repository.findByUuid(uuid);
     const portfolio = await this.portfoliosRepository.findOne(portfolioUuid);
+    if (!portfolio) throw new NotFoundException('Portfolio not found');
     this.checkOwner(user, portfolio);
     await this.repository.deleteByUuidAndPortfolioUuid(portfolioUuid, uuid);
     await this.updatePortfolioState(portfolio);
